Added ignoreAuthBuffer configuration
ignoreAuthBuffer configuration allows a request to modify buffering of
the request on failure. This is particularly useful for handling
application logout in SPAs. The server returns a 401 for logout
requests and this should not be buffered as buffering will mean the
user will be logged out again on successful login